+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - feat(kvstoreentry/delete): Add support for multiple-key deletion using a key prefix. ([#1822](https://github.com/fastly/cli/pull/1822))
 - build(dockerfile-go): add Go Dockerfile alongside the existing Node and Rust ones ([#1828](https://github.com/fastly/cli/pull/1828))
 - feat(compute/deploy): Support 'contentguard' configuration on 'bot_management' product under \[setup.products] ([#1827](https://github.com/fastly/cli/pull/1827))
+- feat(compute): add `install-viceroy` command to pre-install the Viceroy binary ([#1833](https://github.com/fastly/cli/pull/1833))
 
 ### Dependencies: